requirements.txt
python-dotenv
wikipedia
requests
colorama
time
json
sys
re




